feat: :zap: Several Changes - Read desc
This commit is a merge of multiple commits. A product of my lack of xp when it comes to advanced interactive rebasing. I had to fixup several commits since I couldn't apply updates to previous commits properly The changes which are introduced here are: - red outline on wrong otp in web - reset otp process on closing popup in web - persistent sessions for google web and otp mobile - intervals to sync with server - view stats popup